Gesture.py
Hand Gesture is using for ppt Presentation. This project is used for Present a paper without touching and without bluetooth or any other connection. It is very useful for the presenters who are unable to see and who are unable to walk people. This project uses only Hand Signs without any other connection.